refactor(cli): read sandbox IDs through the OpenShell adapter
Move the `openshell sandbox get` call behind the OpenShell identity
adapter and inject it into session detection, so the state module keeps
to parsing and classification. Behavior is unchanged: the lookup is still
memoized per sandbox, still made only when the process list contains a
proxied connection, and still fails soft to SSH-host matching.
